intent_expansion/requirements.txt
sentence-transformers
scikit-learn
numpy
openai
google-generativeai
tqdm
